+++ clj/CHANGELOG.md
@@ -12,7 +12,12 @@ All notable changes to this project will be documented in this file. This change
 - tolerance for balance directives #103
 
 ### Added
+
 - Containerfile for podman/docker #94
+
+### Changed
+
+- appending :directives to filters to pass custom directives is no longer supported
 
 [commit log]: https://github.com/tesujimath/limabean/compare/0.5.1...HEAD